casestudy

+++ case_study/case_study/src/main/java/com/example/case_study/repository/PostTypeRepository.java
@@ -3,5 +3,5 @@ package com.example.case_study.repository;
 import com.example.case_study.model.PostType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostTypeRepository extends JpaRepository<PostType, Long> {
+public interface PostTypeRepository extends JpaRepository<PostType, Integer> {
 }
